Moving settings default and struct.
