+++ crates/apub/src/activity_queue.rs
@@ -22,7 +22,7 @@ use background_jobs::{
 use itertools::Itertools;
 use lemmy_db_queries::DbPool;
 use lemmy_db_schema::source::{community::Community, user::User_};
-use lemmy_utils::{location_info, settings::Settings, LemmyError};
+use lemmy_utils::{location_info, settings::structs::Settings, LemmyError};
 use lemmy_websocket::LemmyContext;
 use log::{debug, warn};
 use reqwest::Client;
