+++ crates/apub/src/fetcher/search.rs
@@ -35,7 +35,7 @@ use lemmy_db_schema::source::{
 use lemmy_db_views::{comment_view::CommentView, post_view::PostView};
 use lemmy_db_views_actor::{community_view::CommunityView, user_view::UserViewSafe};
 use lemmy_structs::{blocking, site::SearchResponse};
-use lemmy_utils::{settings::Settings, LemmyError};
+use lemmy_utils::{settings::structs::Settings, LemmyError};
 use lemmy_websocket::LemmyContext;
 use log::debug;
 use url::Url;
