+++ crates/apub/src/inbox/mod.rs
@@ -19,7 +19,7 @@ use lemmy_db_queries::{
 };
 use lemmy_db_schema::source::{activity::Activity, community::Community, user::User_};
 use lemmy_structs::blocking;
-use lemmy_utils::{location_info, settings::Settings, LemmyError};
+use lemmy_utils::{location_info, settings::structs::Settings, LemmyError};
 use lemmy_websocket::LemmyContext;
 use serde::Serialize;
 use std::fmt::Debug;
