+++ crates/structs/src/lib.rs
@@ -13,7 +13,7 @@ use lemmy_db_schema::source::{
   user::User_,
   user_mention::{UserMention, UserMentionForm},
 };
-use lemmy_utils::{email::send_email, settings::Settings, utils::MentionData, LemmyError};
+use lemmy_utils::{email::send_email, settings::structs::Settings, utils::MentionData, LemmyError};
 use log::error;
 use serde::{Deserialize, Serialize};
 use url::Url;
